// bridge/typescript/src/_generated/program/vm/meta.ts
/* generated by destack_bridge_generate, do not edit */

import { BinaryReader, BinaryWriter, Json, jsonField, jsonInteger, jsonObject } from "../../../protocol/serde.js";

import { ReferenceMetaImpl } from "../../../_impl/program/vm/meta.js";

/** Metadata for reference values. */
export type ReferenceMeta = {
    readonly bits: number;
};

export const ReferenceMeta = {
    ...ReferenceMetaImpl,

    /** Encode this value. */
    encode(writer: BinaryWriter, value: ReferenceMeta): void {
        encodeReferenceMeta(writer, value);
    },

    /** Decode one ReferenceMeta. */
    decode(reader: BinaryReader): ReferenceMeta {
        return decodeReferenceMeta(reader);
    },

    /** Return this value as JSON. */
    toJson(value: ReferenceMeta): Json {
        return toJsonReferenceMeta(value);
    },

    /** Return one ReferenceMeta from one JSON value. */
    fromJson(value: Json): ReferenceMeta {
        return fromJsonReferenceMeta(value);
    },
};

/** Encode one ReferenceMeta. */
export function encodeReferenceMeta(writer: BinaryWriter, value: ReferenceMeta): void {
    writer.writeUnsigned(value.bits);
}

/** Decode one ReferenceMeta. */
export function decodeReferenceMeta(reader: BinaryReader): ReferenceMeta {
    const bits = reader.readNumber();

    return {
        bits,
    };
}

/** Return one JSON value for one ReferenceMeta. */
export function toJsonReferenceMeta(value: ReferenceMeta): Json {
    return {
        bits: value.bits,
    };
}

/** Return one ReferenceMeta from one JSON value. */
export function fromJsonReferenceMeta(value: Json): ReferenceMeta {
    const object = jsonObject(value);

    return {
        bits: jsonInteger(jsonField(object, "bits")),
    };
}